move search logic from Home.vue to SearchResult.vue component
