+++ src/components/layout/TheHeader.vue
@@ -50,80 +50,6 @@
         <div class="sm:flex sm:justify-center md:flex-none">
           <SearchBar />
         </div>
-        <!-- <div
-          class="absolute inset-y-0 right-0 flex items-center pr-2 sm:static sm:inset-auto sm:ml-6 sm:pr-0"
-        > -->
-        <!-- <div
-            class="flex items-center space-x-2 border rounded-md p-1 focus-within:border-primary"
-          >
-            <SearchIcon class="h-4 w-4 text-gray-500" />
-            <input
-              type="text"
-              placeholder="Live search"
-              v-model.trim="inputValue"
-              class="text-sm outline-none placeholder-gray-400 focus:placeholder-gray-300"
-            />
-          </div> -->
-
-        <!-- Profile dropdown -->
-        <!-- <Menu as="div" class="ml-3 relative">
-            <div>
-              <MenuButton
-                class="bg-gray-800 flex text-sm rounded-full focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-800 focus:ring-white"
-              >
-                <span class="sr-only">Open user menu</span>
-                <img
-                  class="h-8 w-8 rounded-full"
-                  src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
-                  alt=""
-                />
-              </MenuButton>
-            </div>
-            <transition
-              enter-active-class="transition ease-out duration-100"
-              enter-from-class="transform opacity-0 scale-95"
-              enter-to-class="transform opacity-100 scale-100"
-              leave-active-class="transition ease-in duration-75"
-              leave-from-class="transform opacity-100 scale-100"
-              leave-to-class="transform opacity-0 scale-95"
-            >
-              <MenuItems
-                class="origin-top-right absolute right-0 mt-2 w-48 rounded-md shadow-lg py-1 bg-white ring-1 ring-black ring-opacity-5 focus:outline-none"
-              >
-                <MenuItem v-slot="{ active }">
-                  <a
-                    href="#"
-                    :class="[
-                      active ? 'bg-gray-100' : '',
-                      'block px-4 py-2 text-sm text-gray-700',
-                    ]"
-                    >Your Profile</a
-                  >
-                </MenuItem>
-                <MenuItem v-slot="{ active }">
-                  <a
-                    href="#"
-                    :class="[
-                      active ? 'bg-gray-100' : '',
-                      'block px-4 py-2 text-sm text-gray-700',
-                    ]"
-                    >Settings</a
-                  >
-                </MenuItem>
-                <MenuItem v-slot="{ active }">
-                  <a
-                    href="#"
-                    :class="[
-                      active ? 'bg-gray-100' : '',
-                      'block px-4 py-2 text-sm text-gray-700',
-                    ]"
-                    >Sign out</a
-                  >
-                </MenuItem>
-              </MenuItems>
-            </transition>
-          </Menu> -->
-        <!-- </div> -->
       </div>
     </div>
 
@@ -160,7 +86,6 @@ export default {
     Disclosure,
     DisclosureButton,
     DisclosurePanel,
-
     MenuIcon,
     XIcon,
   },
